requirements.txt
absl-py==1.2.0
aiofiles==22.1.0
anyio==3.6.1
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
deepface==0.0.75
entrypoints==0.4
executing==0.10.0
fastapi==0.79.1
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==1.12
gast==0.4.0
gdown==4.5.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
httptools==0.4.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mtcnn==0.1.1
nest-asyncio==1.5.5
numpy==1.23.2
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.2.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.4.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
